refactor: move calendar generation to the same file
